Clean up all tests (12 suits, 43 total)
